added 3 decimals instead of 2

+++ database/migrations/2016_05_03_000007_create_purchase_orders.php
@@ -34,7 +34,7 @@ class CreatePurchaseOrders extends Migration
         {
             $table->increments('id');
             $table->integer('status');
-            $table->integer('quantity');
+            $table->decimal('quantity',8,3);
             $table->integer('received')->default(0);
 
             $table->decimal('price'     ,8,2);
